requirements.txt
pyttsx3
SpeechRecognition
pyjokes
pyaudio
pillow
mysql-connector-python